#505421 | change tests because of linter warnings V2

+++ tests/repository_test.go
@@ -21,7 +21,7 @@ func Test_FindById(t *testing.T) {
 	employeeRepository := employee.NewEmployeeRepository(database.ConnectDb())
 	employeeRepository.ExecuteQuery("ALTER SEQUENCE employee_id_seq RESTART WITH 1")
 	_, err2 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName"})
-	if err2 != nil {
+	if err2 == nil {
 		res, err := employeeRepository.FindById(1)
 		if err == nil {
 			assert.Equal(t, "testName", res.Name)
@@ -38,7 +38,7 @@ func Test_FindByIds(t *testing.T) {
 	_, err3 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
 	_, err4 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
 	_, err5 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
-	if err2 != nil && err3 != nil && err4 != nil && err5 != nil {
+	if err2 == nil && err3 == nil && err4 == nil && err5 == nil {
 		ids := make([]int64, 4)
 		ids[0], ids[1], ids[2], ids[3] = 1, 2, 3, 4
 		res, err := employeeRepository.FindByIds(ids)
@@ -52,13 +52,15 @@ func Test_FindByIds(t *testing.T) {
 func Test_FindAll(t *testing.T) {
 	employeeRepository := employee.NewEmployeeRepository(database.ConnectDb())
 	employeeRepository.ExecuteQuery("ALTER SEQUENCE employee_id_seq RESTART WITH 1")
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
-	res, err := employeeRepository.FindAll()
-	if err == nil {
-		assert.Equal(t, 4, len(res))
+	_, err2 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
+	_, err3 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
+	_, err4 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
+	_, err5 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
+	if err2 == nil && err3 == nil && err4 == nil && err5 == nil {
+		res, err := employeeRepository.FindAll()
+		if err == nil {
+			assert.Equal(t, 4, len(res))
+		}
 	}
 	employeeRepository.ExecuteQuery("DELETE FROM employee")
 }
@@ -66,19 +68,21 @@ func Test_FindAll(t *testing.T) {
 func Test_DeleteById(t *testing.T) {
 	employeeRepository := employee.NewEmployeeRepository(database.ConnectDb())
 	employeeRepository.ExecuteQuery("ALTER SEQUENCE employee_id_seq RESTART WITH 1")
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
-	res, err := employeeRepository.FindAll()
-	if err == nil {
-		assert.Equal(t, 4, len(res))
-	}
-	err = employeeRepository.DeleteById(1)
-	if err != nil {
-		res, err = employeeRepository.FindAll()
+	_, err2 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
+	_, err3 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
+	_, err4 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
+	_, err5 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
+	if err2 == nil && err3 == nil && err4 == nil && err5 == nil {
+		res, err := employeeRepository.FindAll()
 		if err == nil {
-			assert.Equal(t, 3, len(res))
+			assert.Equal(t, 4, len(res))
+		}
+		err = employeeRepository.DeleteById(1)
+		if err == nil {
+			res, err = employeeRepository.FindAll()
+			if err == nil {
+				assert.Equal(t, 3, len(res))
+			}
 		}
 	}
 	employeeRepository.ExecuteQuery("DELETE FROM employee")
@@ -87,20 +91,25 @@ func Test_DeleteById(t *testing.T) {
 func Test_DeleteByIds(t *testing.T) {
 	employeeRepository := employee.NewEmployeeRepository(database.ConnectDb())
 	employeeRepository.ExecuteQuery("ALTER SEQUENCE employee_id_seq RESTART WITH 1")
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
-	employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
-	res, err := employeeRepository.FindAll()
-	if err == nil {
-		assert.Equal(t, 4, len(res))
-	}
-	ids := make([]int64, 2)
-	ids[0], ids[1] = 1, 2
-	err = employeeRepository.DeleteByIds(ids)
-	res, err = employeeRepository.FindAll()
-	if err == nil {
-		assert.Equal(t, 2, len(res))
+	_, err2 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName1"})
+	_, err3 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName2"})
+	_, err4 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName3"})
+	_, err5 := employeeRepository.Save(employee.EmployeeEntity{Name: "testName4"})
+	if err2 == nil && err3 == nil && err4 == nil && err5 == nil {
+		res, err := employeeRepository.FindAll()
+		if err == nil {
+			assert.Equal(t, 4, len(res))
+		}
+		ids := make([]int64, 2)
+		ids[0], ids[1] = 1, 2
+		err = employeeRepository.DeleteByIds(ids)
+		if err == nil {
+			res, err = employeeRepository.FindAll()
+		}
+
+		if err == nil {
+			assert.Equal(t, 2, len(res))
+		}
 	}
 	employeeRepository.ExecuteQuery("DELETE FROM employee")
 }
